+++ src_docs/components/simple-list.component.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import faker from 'faker';
 import List from '../../src/index';
 
 export default class SimpleList extends React.PureComponent {
@@ -8,13 +7,6 @@ export default class SimpleList extends React.PureComponent {
     this.state = {
       selectedItems: [],
     };
-    this.items = [];
-    for (let i = 1; i <= 50; i += 1) {
-      this.items.push({
-        id: i,
-        value: faker.name.findName(),
-      });
-    }
   }
 
   onSelectedChange = (selectedItems) => {
@@ -27,7 +19,6 @@ export default class SimpleList extends React.PureComponent {
     } = this.state;
     return (
       <List
-        items={this.items}
         selectedItems={selectedItems}
         onSelectedChange={this.onSelectedChange}
         {...this.props}